addCaster, editCaster, deleteCaster for admin (editMovie n editCaster msh blm fix)

+++ admin.php
@@ -14,7 +14,10 @@ if ($_SESSION['role'] !== 'Admin') {
 
 $user = query("SELECT * FROM users WHERE role = 'User'");
 
-$movie = query("SELECT * FROM movies")
+$movie = query("SELECT * FROM movies");
+
+$caster = query("SELECT casters.*, movies.title AS movie_title FROM casters LEFT JOIN movies ON casters.id_movies = movies.id");
+
 ?>
 
 <!DOCTYPE html>
@@ -48,7 +51,7 @@ $movie = query("SELECT * FROM movies")
                 <td><?= $row['firstName']; ?></td>
                 <td><?= $row['lastName']; ?></td>
                 <td>
-                    <a href="user/editUser.php?id=<?= $row['id']; ?>">Edit</a>|
+                    <a href="user/editUser.php?id=<?= $row['id']; ?>">Edit</a> |
                     <a href="user/deleteUser.php?id=<?= $row['id']; ?>" onclick="return confirm('Are you sure?')">Delete</a>
                 </td>
             </tr>
@@ -57,7 +60,7 @@ $movie = query("SELECT * FROM movies")
     </table><br><br>
 
     <h3>Table Movies</h3>
-    <a href="movie/addMovie.php">Add Movie</a><br><br>
+    <a href="movie/addMovie.php">Add Data</a><br><br>
     <table>
         <tr>
             <th>No.</th>
@@ -78,11 +81,40 @@ $movie = query("SELECT * FROM movies")
                 <td><?= $row['year']; ?></td>
                 <td><?= $row['director']; ?></td>
                 <td>
-                    <img src="./assets/upload/images/<?= $row['picture'] ?>" width="75">
+                    <img src="assets/upload/images/<?= $row['picture'] ?>" width="75">
                 </td>
                 <td>
-                    <a href="movie/editMovie.php?id=<?= $row['id']; ?>">Edit</a>|
-                    <a href="movie/deleteMovie.php?id=<?= $row['id']; ?>">Delete</a>
+                    <a href="movie/editMovie.php?id=<?= $row['id']; ?>">Edit</a> |
+                    <a href="movie/deleteMovie.php?id=<?= $row['id']; ?>" onclick="return confirm('Are you sure?')">Delete</a>
+                </td>
+            </tr>
+            <?php $i++; ?>
+        <?php endforeach; ?>
+    </table><br><br>
+
+    <h3>Table Casters</h3>
+    <a href="caster/addCaster.php">Add Data</a><br><br>
+    <table>
+        <tr>
+            <th>No.</th>
+            <th>Name</th>
+            <th>Movie Name</th>
+            <th>Picture</th>
+            <th>Action</th>
+        </tr>
+
+        <?php $i = 1; ?>
+        <?php foreach ($caster as $row) : ?>
+            <tr>
+                <td><?= $i; ?></td>
+                <td><?= $row['name']; ?></td>
+                <td><?= $row['movie_title']; ?></td>
+                <td>
+                    <img src="assets/upload/images/<?= $row['picture'] ?>" width="75">
+                </td>
+                <td>
+                    <a href="caster/editCaster.php?id=<?= $row['id']; ?>">Edit</a> |
+                    <a href="caster/deleteCaster.php?id=<?= $row['id']; ?>" onclick="return confirm('Are you sure?')">Delete</a>
                 </td>
             </tr>
             <?php $i++; ?>
